# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos/{publicId}/status": {
      "get": {
        "description": "Returns the video's processing status and metadata. Only the owning channel's user may query it.",
        "operationId": "VideosController_getStatus",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "Uakgb_J5m9j",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Video status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoStatusResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Video does not belong to the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get video processing status",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/stream-url": {
      "get": {
        "description": "Issues a time-limited presigned GET URL direct from object storage, supporting HTTP Range requests. Requires the video to be ready.",
        "operationId": "VideosController_getStreamUrl",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "Uakgb_J5m9j",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presigned streaming URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Get a presigned streaming URL",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/download-url": {
      "get": {
        "description": "Issues a time-limited presigned GET URL with an attachment content-disposition, direct from object storage. Requires the video to be ready.",
        "operationId": "VideosController_getDownloadUrl",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "Uakgb_J5m9j",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presigned download URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Get a presigned download URL",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VideoStatusResponseDto": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "example": "Uakgb_J5m9j"
          },
          "title": {
            "type": "string",
            "example": "my-video.mp4"
          },
          "processingStatus": {
            "type": "string",
            "enum": [
              "uploading",
              "processing",
              "ready",
              "failed"
            ],
            "example": "ready"
          },
          "durationSeconds": {
            "type": "object",
            "nullable": true,
            "example": 12.5
          },
          "width": {
            "type": "object",
            "nullable": true,
            "example": 1920
          },
          "height": {
            "type": "object",
            "nullable": true,
            "example": 1080
          },
          "thumbnailUrl": {
            "type": "object",
            "nullable": true,
            "example": "https://minio.example/thumbnails/..."
          }
        },
        "required": [
          "publicId",
          "title",
          "processingStatus",
          "durationSeconds",
          "width",
          "height",
          "thumbnailUrl"
        ]
      },
      "PresignedUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://minio.example/videos/..."
          },
          "expiresInSeconds": {
            "type": "number",
            "example": 3600
          }
        },
        "required": [
          "url",
          "expiresInSeconds"
        ]
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}